refactor(agent): group modelHandlers by provider sub-domain
Organize the flat modelHandlers grab-bag (the model-provider bounded
context) into one directory per provider family so the layout mirrors the
ubiquitous language:

- anthropic/   handler + Anthropic-only helpers
- openai/      handler, Responses handler, OpenAI helpers, and the
               OpenAI-compatible providers (DashScope/DeepSeek/GLM/Kimi/
               MiniMax/XAI) that subclass ModelHandlerOpenAI
- google/      GoogleGenAI handler
- openrouter/  OpenRouter native handler

Genuinely cross-provider code stays at the root (ModelHandler base,
toolConversion, contextManagementConstants, modelHandlerValidation) along
with the shared support/, types/, and utils/ subdirs. Added a README
documenting the boundaries.

Pure move + import-path rewrite; no barrel, no re-export shims. Importers
use @agent/modelHandlers/<provider>/<File>.
Verified: typecheck, lint (0 errors), and test:kernel (1378 tests) pass.
