+++ src/agent/modelHandlers/openai/openAIMessageUtils.ts
