+++ src/test/agent/modelHandlers/ModelHandlerOpenAIResponseBackground.test.ts
@@ -11,7 +11,7 @@ import {
 // Local imports - agent
 import { AgentCategory } from '@agent/core/definition/AgentDataclass';
 import * as configModule from '@agent/core/config';
-import { ModelHandlerOpenAIResponse } from '@agent/modelHandlers/modelHandlerOpenAIResponse';
+import { ModelHandlerOpenAIResponse } from '@agent/modelHandlers/openai/modelHandlerOpenAIResponse';
 
 class UnsupportedBackgroundHandler extends ModelHandlerOpenAIResponse {
   protected override backgroundModeSupported = false;
